requirements.txt
shapely
lap
streamlit
streamlit-extras
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
moviepy==1.0.3
imageio==2.34.1
imageio-ffmpeg==0.4.9
opencv-python-headless==4.10.0.84
numpy==2.0.2
ultralytics==8.3.118
pandas==2.2.3
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
pillow==11.0.0
tqdm==4.67.1
matplotlib==3.10.0
ultralytics-thop==2.0.14


